backend/fastapi/requirements.txt
alembic==1.13.3
asyncpg==0.30.0
fastapi==0.115.4
greenlet==3.1.1
httpx==0.27.2
passlib[bcrypt]==1.7.4
pydantic-settings==2.6.1
pydantic[email]==2.9.2
pyjwt==2.9.0
pytest==8.3.3
python-dotenv==1.0.0
pytest-asyncio==0.24.0
pytest-cov==6.0.0
python-slugify==8.0.4
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
starlette==0.41.2
structlog==24.4.0
uvicorn==0.32.0